Some explanations about filtering type in Compiler
State handle different Stack of StateAttribute and TextureAttribute.
StateAttribute are mostly things like BlendFunc, Depth, ... user can create
custom StateAttribute IF it does not overlap with gl state in another one.
Example if you create a custom StateAttribute that manipulate BlendFunc, or a
Texture, then the attribute type should be keep in order that the gl state is
not used in different stacks.